+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2025-01-09
+- #2183 adds a way to collect and store metrics from influxDB. This is not a breaking change for customers of the SDK.
+
 ## 2025-01-08
 - #2170 increases the suggested value for `DEFERRED_SLOTS_COUNT` to 50. This value is more appropriate than the previous value (5) for testing, but still too low for production use. A value in the order of 1000 is more appropriate for production.